Settled ammo counters location, added weapon slot groups to tank encode/decode, added reload command

+++ Assets/Scripts/Models/Tank/Weapons/BaseWeapon.cs
@@ -145,7 +145,7 @@ namespace TankArena.Models.Weapons
         public bool isReloading;
         public bool isShooting;
         protected float currentReloadTimer;
-        public float currentClipSize;
+        public int currentClipSize;
         protected float maxShotDelay;
         
         protected BaseWeaponController weaponController;
